requirements.txt
Flask
flask_cors
kafka-python
Flask-SocketIO
python-socketio
python-engineio
numpy
opencv-python-headless
imutils
pyopenssl
Pillow
scipy
scikit-learn==0.23.2
mlxtend
dlib
imageio
scikit-image
pandas
eventlet